Anxiety-Detection-from-Speech/wave2vecprob/requirements.txt
# ============================================
# Speech Anxiety Detection — Dependencies
# ============================================

# Core ML
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.38.0
datasets>=2.0.0
peft>=0.10.0

# Audio Processing
opensmile>=2.4.0
librosa>=0.10.0
soundfile>=0.12.0
audiomentations>=0.33.0

# Training & Monitoring
scikit-learn>=1.3.0
tensorboard>=2.14.0
accelerate>=0.27.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
xgboost>=1.7.0

# Inference & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydub>=0.25.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
